requirements.txt
joblib==1.3.2
numpy==1.26.2
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
